feat: configure dev env and layout
