+++ package.json
@@ -31,6 +31,7 @@
     "eslint-plugin-prettier": "5.1.3",
     "eslint-plugin-react": "7.34.3",
     "eslint-plugin-react-hooks": "4.6.2",
+    "dotenv": "16.4.7",
     "nx": "21.5.3",
     "prettier": "3.4.2",
     "typescript": "5.5.4"
