+++ src/microfrontends/operations-reports/.env
@@ -1,0 +1,6 @@
+MICROFRONT_PORT=4400
+MICROFRONT_HOST=0.0.0.0
+MICROFRONT_PUBLIC_URL=http://localhost:4401
+CLIENT_PORT=4401
+CLIENT_HOST=0.0.0.0
+SHELL_URL=http://localhost:4300
