+++ src/microfrontends/operations-reports/server/project.json
@@ -15,7 +15,7 @@
       "executor": "nx:run-commands",
       "options": {
         "cwd": "src/microfrontends/operations-reports/server",
-        "command": "npm run start"
+        "command": "npm run dev"
       }
     }
   },
