+++ src/microfrontends/users-and-roles/.env
@@ -1,0 +1,6 @@
+MICROFRONT_PORT=4402
+MICROFRONT_HOST=0.0.0.0
+MICROFRONT_PUBLIC_URL=http://localhost:4403
+CLIENT_PORT=4403
+CLIENT_HOST=0.0.0.0
+SHELL_URL=http://localhost:4300
