+++ src/shell-app/.env
@@ -1,0 +1,4 @@
+SHELL_PORT=4300
+SHELL_HOST=0.0.0.0
+CLIENT_PORT=4301
+CLIENT_HOST=0.0.0.0
