+++ src/shell-app/server/package.json
@@ -4,9 +4,14 @@
   "private": true,
   "scripts": {
     "build": "node scripts/build.js",
-    "start": "node shell-server.js"
+    "start": "node shell-server.js",
+    "dev": "nodemon --watch shell-server.js --watch data --ext js,json --exec node shell-server.js"
   },
   "dependencies": {
-    "express": "4.19.2"
+    "express": "4.19.2",
+    "dotenv": "16.4.7"
+  },
+  "devDependencies": {
+    "nodemon": "3.1.7"
   }
 }
